+++ src/aw.js
@@ -16,7 +16,7 @@ function aw (...args) {
 
 function _aw (fn, options = {}) {
   return async function (...args) {
-    const opts = { ...options, context: options.context || this }
+    const opts = { context: this, injectCallback: true, ...options }
     try {
       var { type, val } = await wrap(fn, args, opts)
       if (type === 'cb' && Array.isArray(val)) return [null, ...val]
@@ -29,7 +29,9 @@ function _aw (fn, options = {}) {
 
 function wrap (fn, args, options) {
   return new Promise(function (resolve, reject) {
-    args.push(function injectedAWCallbackArg (err, ...cbArgs) { err == null ? resolve({ type: 'cb', val: cbArgs }) : reject(err) })
+    if (options.injectCallback) {
+      args.push(function injectedAWCallbackArg (err, ...cbArgs) { err == null ? resolve({ type: 'cb', val: cbArgs }) : reject(err) })
+    }
 
     let p = fn.apply(options.context, args)
     if (p instanceof Promise) {